#SDSV-5 Improve vertical layout code
